refactor: app/_common을 common으로
