+++ tailwind.config.ts
@@ -2,7 +2,7 @@ import type { Config } from "tailwindcss";
 
 const config: Config = {
   darkMode: "class",
-  content: ["./app/**/*.{js,ts,jsx,tsx,mdx}"],
+  content: ["./app/**/*.{ts,tsx}", "./common/**/*.{ts,tsx}"],
   theme: {
     extend: {
       fontFamily: {
